backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.8.3
chromadb==0.5.23
python-dotenv==1.0.1
httpx==0.27.0
